third_party/blink/web_tests/media/W3C/video/readyState/readyState_during_loadeddata-expected.txt

video.readyState should be >= HAVE_CURRENT_DATA during loadeddata event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS true is true
PASS successfullyParsed is true

TEST COMPLETE


spec reference
